Refactor - new name for gem, namespaces, remove unused code.
