+++ Gemfile
@@ -17,7 +17,7 @@ gem "pg"
 gem "plek"
 gem "sassc-rails"
 gem "sentry-sidekiq"
-gem "site-improve-api-client", github: "alphagov/site-improve-api-client", branch: "fix-difficulty"
+gem "siteimprove_api_client", github: "alphagov/siteimprove_api_client"
 gem "uglifier"
 
 group :test do
