--- app/services/fetch_site_improve_summary.rb
@@ -1,17 +1,0 @@
-class FetchSiteImproveSummary
-  include SiteImproveApiClient
-
-  def self.call(**args)
-    new(**args).call
-  end
-
-  def initialize(url:)
-    @url = url
-  end
-
-  def call
-    Rails.cache.fetch("site-improve-summary-for/#{@url}", expires_in: 1.hour) do
-      content_api.sites_site_id_content_pages_page_id_get(site_id, page_id_for_url(url: @url))
-    end
-  end
-end
