--- app/views/metrics/_accessibility_issues.erb
@@ -1,26 +1,0 @@
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds content-metrics" data-gtm-id="content-metrics">
-    <h3 class="content-metrics__header govuk-heading-m">
-
-      <%= local_assigns[:title] %>
-    </h3>
-    <% if local_assigns[:show_link] %>
-      <p><a href="<%= local_assigns[:accessibility_issues].link %>"><%= t("metrics.siteimprove.accessibility_report_link") %></a></p>
-    <% end %>
-        <% local_assigns[:accessibility_issues].issue_list.each do |issue| %>
-          <p>
-            <%= issue.help.title%>
-          </p>
-
-          <%= render "govuk_publishing_components/components/details", {
-            title: t("metrics.siteimprove.more_issue_detail")
-          } do %>
-            <%= raw(issue.help.description) %>
-          <% end %>
-
-        <% end%>
-
-  </div>
-</div>
-
-
